feat(dashboard): impl link parsing from err messages
