+++ package.json
@@ -162,6 +162,7 @@
     "jose": "~5.10.0",
     "jwks-rsa": "^3.1.0",
     "kafkajs": "^2.2.4",
+    "linkify-it": "^5.0.1",
     "lucide-react": "0.561.0",
     "motion": "^12.23.0",
     "msw": "^2.12.2",
@@ -266,6 +267,7 @@
     "@types/dockerode": "^3.3.34",
     "@types/ejs": "^3.1.5",
     "@types/jest": "30.0.0",
+    "@types/linkify-it": "^5.0.0",
     "@types/multer": "^1.4.12",
     "@types/node": "^22.13.4",
     "@types/nodemailer": "^8.0.0",
